refactor(icon): add src and types folders
