--- packages/icon/Icon.d.ts
@@ -1,8 +1,0 @@
-export interface IconProps extends React.HTMLAttributes<HTMLElement> {
-    name: string;
-    size?: string;
-}
-
-declare const Icon: React.FunctionComponent<IconProps>;
-
-export default Icon;
