+++ packages/icon/index.d.ts
@@ -1,1 +1,1 @@
-export {default} from './Icon';
+export { default } from './types/Icon';
