+++ packages/icon/package.json
@@ -28,6 +28,7 @@
     "react": ">=16.3.0"
   },
   "dependencies": {
-    "classnames": "^2.3.1"
+    "classnames": "^2.3.1",
+    "prop-types": "^15.7.2"
   }
 }
