+++ packages/icon/types/Icon.d.ts
@@ -1,0 +1,8 @@
+export interface IconProps extends React.HTMLAttributes<HTMLElement> {
+  name: string;
+  size?: string;
+}
+
+declare const Icon: React.FC<IconProps>;
+
+export default Icon;
